+++ apps/frontend/src/modules/users/slices/users.ts
@@ -1,10 +1,9 @@
-import { getAuthenticatedUser, loadAll } from "./actions.js";
+import { getAuthenticatedUser } from "./actions.js";
 import { actions } from "./users.slice.js";
 
 const allActions = {
 	...actions,
 	getAuthenticatedUser,
-	loadAll,
 };
 
 export { allActions as actions };