+++ packages/shared/src/index.ts
@@ -33,7 +33,6 @@ export { AuthApiPath } from "./modules/auth/auth.js";
 export { QuizApiPath } from "./modules/quiz/quiz.js";
 export {
 	type UserDto,
-	type UserGetAllItemResponseDto,
 	type UserGetAllResponseDto,
 	UsersApiPath,
 	type UserSignInRequestDto,