+++ apps/backend/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [1.7.0](https://github.com/BinaryStudioAcademy/bsa-2024-bebalance/compare/backend-v1.6.0...backend-v1.7.0) (2024-08-25)
+
+
+### Features
+
+* **backend:** onboarding questions and answers bb-133 ([#152](https://github.com/BinaryStudioAcademy/bsa-2024-bebalance/issues/152)) ([444ba83](https://github.com/BinaryStudioAcademy/bsa-2024-bebalance/commit/444ba83f6ceaf088a36a146ace3809a9924d83d9))
+
+## [1.6.0](https://github.com/BinaryStudioAcademy/bsa-2024-bebalance/compare/backend-v1.5.0...backend-v1.6.0) (2024-08-24)
+
+
+### Features
+
+* **backend/frontend:** clean up bb-69 ([#117](https://github.com/BinaryStudioAcademy/bsa-2024-bebalance/issues/117)) ([955363d](https://github.com/BinaryStudioAcademy/bsa-2024-bebalance/commit/955363d521e636a287ec2f69b002dd6b981dce33))
+
 ## [1.5.0](https://github.com/BinaryStudioAcademy/bsa-2024-bebalance/compare/backend-v1.4.4...backend-v1.5.0) (2024-08-23)